finished basic structure
