+++ .gitignore
@@ -24,3 +24,4 @@ Cargo.lock
 
 /target
 .DS_Store
+todo.txt
