+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod validation;
 mod codelist;
 mod errors;
+mod codelist_type;
