+++ src/validation/mod.rs
@@ -1,3 +1,3 @@
-pub mod icd10_validator;
 pub mod opcs_validator;
 pub mod snomed_validator;
+pub mod icd10_validator;
