+++ src/validation/opcs_validator.rs
@@ -1,13 +1,19 @@
 use crate::codelist::CodeList;
 
 pub trait OPCSValidator {
-    fn validate(&self, code: &str) -> bool;
+    fn validate_code(&self, code: &str) -> bool; // for 1 code
+    fn validate_all_code(&self);
 }
 
 // Implement the `OPCSValidator` trait for `CodeList`
 impl OPCSValidator for CodeList {
-    fn validate(&self, code: &str) -> bool {
+    fn validate_code(&self, code: &str) -> bool {
         //TODO
         true
     }
+    fn validate_all_code(&self) {
+
+    }
 }
+
+//TODO: tests
